extracted FilterByContent
